Commit Final Finaloso de Finales

+++ Program.cs
@@ -34,7 +34,7 @@
                     Console.WriteLine("El resultado de la multiplicacion de " + num1 + " por " + num2 + " es: " + multiplicacion);
                     break;
                 case 4:
-                    division = num1 / num2;
+                    division = (float)(num1 / num2);
                     Console.WriteLine("$El resultado de la división entre {0} y {1} es {2}", num1, num2, division);
                     break;
                 default:
